Some reshuffling of code to make it cleaner / maintainable.
